[wpa_supplicant] Re-enable WEP. JB#60571
It's no good, but supporting it here doesn't cost much, and the
earlier disabling didn't seem all concious decision, rather
a side-effect on both upstream changing the default value and
at the same time packaging changing the build config handling.

For comparison WEP is still enabled for Fedora and Android 13
I have for testing.

For Sailfish the proper removal should be done first on the UI level,
which hasn't been done so far.