[wpa_supplicant] Update to 2.11. Add patches from distros. Fixes JB#62420
Update to latest wpa_supplicant version 2.11.

Drop one old patch that reverted a change bug does not seem to be
necessary anymore on old devices, the comment was missing from commit
and was in JB#44048 comment 13: "Added a revert for a patch which
prevented this package to work on older < 4.0 kernels"

Apply the relevant patches from Fedora/Debian and SuSe (in this order).
Few of the patches make debugging wpa_supplicant easier.