[wpa_supplicant] Revert to older, v.2.9, build_config. Fixes JB#62505
The build_config in v.2.10 broke AP mode on some devices, the network
was not visible. It seems that the old build_config makes this work.
This simply reverts the changes made and further work is needed to
figure out what exactly breaks the AP mode.

Drop the build config options that were removed in 2.10 in comparison to
the config that was in 2.9:
 CONFIG_IEEE80211N
 CONFIG_IEEE80211W
 CONFIG_PEERKEY
 CONFIG_CTRL_IFACE_DBUS