[wpa_supplicant] Enable timeout for sched scan, add scan stop notify. Fixes JB#62571
Add two patches to 1) enable timeout for scheduled scan and 2) to also
notify always when the scheduled scan is stopped. The scheduled scan
timeout is the same as with the normal scan requests, providing results
(with same type as normal scan as there seems to be none for this).
Notify is added because ConnMan, via libgsupplicant, relies on the
scanning status in order to initiate its own scan.